SunSensors simulator
Major refactoring and update of the earlier simulator, with intention to
extend the support to more general class of Sun sensors instead of
focusing only on profile sensors. Support for PSD sensors is added.

The simulator now heavily relies on the Meshes.jl library. Because the
Meshes.jl is well integrated with Unitful.jl, the simulator now also
adopted usage of units.

Code was properly documented and a number of other minor to medium
updates are done.